# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TripJar API - collaborative group savings platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jars/{jarId}/contributions:
    get:
      operationId: listContributions
      tags: [contributions]
      summary: List contributions for a jar
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contribution"
    post:
      operationId: createContribution
      tags: [contributions]
      summary: Add a contribution to a jar
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContributionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"
  /jars/{jarId}/contributions/{contributionId}/reverse:
    post:
      operationId: reverseContribution
      tags: [contributions]
      summary: Reverse a contribution
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
        - name: contributionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReverseContributionRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"

  # SCHEDULES
...
  /jars/{jarId}/commitments:
    get:
      operationId: listCommitmentRequests
      tags: [commitments]
      summary: List commitment requests for a jar
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommitmentRequest"
    post:
      operationId: createCommitmentRequest
      tags: [commitments]
      summary: Create a commitment request
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommitmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitmentRequest"
  /jars/{jarId}/commitments/{commitmentId}/vote:
    post:
      operationId: voteOnCommitment
      tags: [commitments]
      summary: Vote on a commitment request
      parameters:
        - name: jarId
          in: path
          required: true
          schema:
            type: string
        - name: commitmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitmentVoteRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitmentRequest"

  # AGREEMENTS
...
components:
  schemas:
...
    Contribution:
      type: object
      properties:
        id:
          type: string
        jarId:
          type: string
        memberId:
          type: string
        amountCents:
          type: integer
        contributionDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, completed, failed, reversed, refunded, simulated]
        sourceType:
          type: string
          enum: [manual, scheduled, simulated]
        milestoneId:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        memberName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, jarId, memberId, amountCents, contributionDate, status, sourceType, createdAt]
    CreateContributionRequest:
      type: object
      properties:
        amountCents:
          type: integer
          minimum: 1
        contributionDate:
          type: string
          format: date
        milestoneId:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required: [amountCents]
    ReverseContributionRequest:
      type: object
      properties:
        reason:
          type: string

    # SCHEDULES
...
    CommitmentRequest:
      type: object
      properties:
        id:
          type: string
        jarId:
          type: string
        milestoneId:
          type: string
          nullable: true
        amountCents:
          type: integer
        purpose:
          type: string
        vendorName:
          type: string
          nullable: true
        paymentDeadline:
          type: string
          format: date
          nullable: true
        requiredThreshold:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, expired, cancelled]
        approvedCount:
          type: integer
        rejectedCount:
          type: integer
        pendingCount:
          type: integer
        approvalPercentage:
          type: number
        votes:
          type: array
          items:
            $ref: "#/components/schemas/CommitmentVote"
        createdAt:
          type: string
          format: date-time
      required: [id, jarId, amountCents, purpose, requiredThreshold, status, approvedCount, rejectedCount, pendingCount, approvalPercentage, votes, createdAt]
    CreateCommitmentRequest:
      type: object
      properties:
        milestoneId:
          type: string
          nullable: true
        amountCents:
          type: integer
          minimum: 1
        purpose:
          type: string
        vendorName:
          type: string
          nullable: true
        paymentDeadline:
          type: string
          format: date
          nullable: true
        requiredThreshold:
          type: number
          default: 0.67
        note:
          type: string
          nullable: true
      required: [amountCents, purpose]
    CommitmentVote:
      type: object
      properties:
        id:
          type: string
        commitmentRequestId:
          type: string
        memberId:
          type: string
        vote:
          type: string
          enum: [approve, reject, clarification]
        note:
          type: string
          nullable: true
        memberName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, commitmentRequestId, memberId, vote, createdAt]
    CommitmentVoteRequest:
      type: object
      properties:
        vote:
          type: string
          enum: [approve, reject, clarification]
        note:
          type: string
          nullable: true
      required: [vote]

    # AGREEMENTS
...